Added more tests and a lot of small changes

+++ src/main/java/com/github/f4b6a3/uuid/distrib/Distributor.java
@@ -1,5 +1,5 @@
 package com.github.f4b6a3.uuid.distrib;
 
 public interface Distributor {
-	int handOut();
+	long handOut();
 }
